build.gradle: version 0.7.0 -> 0.7.1
Actors:
setSizeByWidth(Actor actor, int width) -> setSizeByWidth(Actor actor, float width)
setSizeByHeight(Actor actor, int height) -> setSizeByHeight(Actor actor, float height)

ClickListeners:
newDownUpColorChange -> newColorChange
addDownUpColorChange -> addColorChange

+++ src/main/java/org/ryuu/gdx/scenes/scene2d/utils/Actors.java
@@ -53,12 +53,12 @@ public class Actors {
         setSizeIsotropic(self, reference, 1);
     }
 
-    public static void setSizeByWidth(Actor actor, int width) {
+    public static void setSizeByWidth(Actor actor, float width) {
         float scale = width / actor.getWidth();
         actor.setSize(actor.getWidth() * scale, actor.getHeight() * scale);
     }
 
-    public static void setSizeByHeight(Actor actor, int height) {
+    public static void setSizeByHeight(Actor actor, float height) {
         float scale = height / actor.getHeight();
         actor.setSize(actor.getWidth() * scale, actor.getHeight() * scale);
     }
